requirements.txt
Flask==3.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0

